# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AutoServis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with the application password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the application password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /gdpr/search:
    get:
      operationId: gdprSearch
      tags: [gdpr]
      summary: Search personal data by name, phone, e-mail or license plate
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching vehicles holding personal data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GdprSearchResult"
  /gdpr/export/{vehicleId}:
    get:
      operationId: gdprExportVehicle
      tags: [gdpr]
      summary: Export all data tied to a vehicle (data subject access request)
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full export document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GdprExport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gdpr/anonymize/{vehicleId}:
    post:
      operationId: gdprAnonymizeVehicle
      tags: [gdpr]
      summary: Remove personal data but keep technical service history
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anonymized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GdprActionResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gdpr/vehicle/{vehicleId}:
    delete:
      operationId: gdprDeleteVehicle
      tags: [gdpr]
      summary: Permanently delete a vehicle and all linked records
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GdprActionResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gdpr/consent/{vehicleId}:
    put:
      operationId: setVehicleConsent
      tags: [gdpr]
      summary: Record or withdraw the owner's data-processing consent
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetConsentInput"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /gdpr/audit-log:
    get:
      operationId: getAuditLog
      tags: [gdpr]
      summary: List recent audit-log entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit entries (most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
    Vehicle:
      type: object
      required: [id, licensePlate, make, model, ownerType, createdAt]
      properties:
        id:
          type: integer
        licensePlate:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: ["integer", "null"]
        color:
          type: ["string", "null"]
        vin:
          type: ["string", "null"]
        engineDisplacement:
          type: ["integer", "null"]
        registrationDate:
          type: ["string", "null"]
        transmission:
          type: ["string", "null"]
          enum: [manual, automatic, null]
        ownerType:
          type: string
          enum: [private, company]
        ownerName:
          type: ["string", "null"]
        ownerAddress:
          type: ["string", "null"]
        ownerIco:
          type: ["string", "null"]
        ownerDic:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        consentGivenAt:
          type: ["string", "null"]
        consentNote:
          type: ["string", "null"]
        currentKm:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        stkValidUntil:
          type: ["string", "null"]
        lastOilChangeKm:
          type: ["integer", "null"]
        lastOilChangeDate:
          type: ["string", "null"]
        lastBrakesDate:
          type: ["string", "null"]
        lastTimingDate:
          type: ["string", "null"]
        lastTransmissionOilDate:
          type: ["string", "null"]
        lastTransmissionOilKm:
          type: ["integer", "null"]
        lastBrakeFluidDate:
          type: ["string", "null"]
        oilChangeIntervalKm:
          type: ["integer", "null"]
        oilChangeIntervalMonths:
          type: ["integer", "null"]
        transmissionOilIntervalKm:
          type: ["integer", "null"]
        transmissionOilIntervalMonths:
          type: ["integer", "null"]
        brakesIntervalMonths:
          type: ["integer", "null"]
        timingIntervalKm:
          type: ["integer", "null"]
        timingIntervalMonths:
          type: ["integer", "null"]
        brakeFluidIntervalMonths:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    ServiceRecord:
      type: object
      required: [id, vehicleId, date, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        date:
          type: string
        km:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        oilChanged:
          type: boolean
        brakesServiced:
          type: boolean
        timingServiced:
          type: boolean
        transmissionOilChanged:
          type: boolean
        brakeFluidChanged:
          type: boolean
        stkPassed:
          type: boolean
        otherWork:
          type: ["string", "null"]
        technician:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    WorkOrder:
      type: object
      required: [id, licensePlate, status, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: ["integer", "null"]
        licensePlate:
          type: string
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, waiting_parts, needs_return, completed]
        km:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        oilChange:
          type: boolean
        transmissionOil:
          type: boolean
        brakes:
          type: boolean
        timing:
          type: boolean
        airFilter:
          type: boolean
        cabinFilter:
          type: boolean
        stk:
          type: boolean
        tireChange:
          type: boolean
        diagnostics:
          type: boolean
        lightsCheck:
          type: boolean
        brakeFluid:
          type: boolean
        frontAxleCheck:
          type: boolean
        rearAxleCheck:
          type: boolean
        frontShocksCheck:
          type: boolean
        rearShocksCheck:
          type: boolean
        geometry:
          type: boolean
        headlightAlignment:
          type: boolean
        serviceDate:
          type: ["string", "null"]
          description: Date when service was performed (for backdated/historical orders)
        otherWork:
          type: ["string", "null"]
        otherServices:
          type: ["string", "null"]
        laborHours:
          type: ["string", "null"]
          description: Number of labor hours (decimal as string, e.g. "2.5")
        laborPrice:
          type: ["integer", "null"]
          description: Labor price in CZK
        notes:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
...
    Photo:
      type: object
      required: [id, workOrderId, url, createdAt]
      properties:
        id:
          type: integer
        workOrderId:
          type: integer
        url:
          type: string
        filename:
          type: string
        createdAt:
          type: string
          format: date-time
    Appointment:
      type: object
      required: [id, scheduledDate, status, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: ["integer", "null"]
        licensePlate:
          type: ["string", "null"]
        scheduledDate:
          type: string
        scheduledTime:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planned, done, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    GdprVehicleMatch:
      type: object
      required: [id, licensePlate, ownerType, serviceRecordCount, workOrderCount, appointmentCount]
      properties:
        id:
          type: integer
        licensePlate:
          type: string
        ownerType:
          type: string
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        consentGivenAt:
          type: ["string", "null"]
        serviceRecordCount:
          type: integer
        workOrderCount:
          type: integer
        appointmentCount:
          type: integer
    GdprSearchResult:
      type: object
      required: [vehicles]
      properties:
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/GdprVehicleMatch"
    GdprExport:
      type: object
      required: [exportedAt, vehicle, serviceRecords, workOrders, appointments]
      properties:
        exportedAt:
          type: string
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        serviceRecords:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRecord"
        workOrders:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrder"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    SetConsentInput:
      type: object
      required: [given]
      properties:
        given:
          type: boolean
        note:
          type: ["string", "null"]
    GdprActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    AuditLogEntry:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: integer
        action:
          type: string
        entity:
          type: ["string", "null"]
        entityId:
          type: ["string", "null"]
        detail:
          type: ["string", "null"]
        createdAt:
          type: string