# lib/api-zod/src/generated/types/gdprExport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoServis API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { ServiceRecord } from './serviceRecord';
import type { Vehicle } from './vehicle';
import type { WorkOrder } from './workOrder';

export interface GdprExport {
  exportedAt: string;
  vehicle: Vehicle;
  serviceRecords: ServiceRecord[];
  workOrders: WorkOrder[];
  appointments: Appointment[];
}