# lib/api-zod/src/generated/types/setConsentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoServis API
 * OpenAPI spec version: 0.1.0
 */

export interface SetConsentInput {
  given: boolean;
  /** @nullable */
  note?: string | null;
}