# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoServis API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleOwnerType } from './vehicleOwnerType';
import type { VehicleTransmission } from './vehicleTransmission';

export interface Vehicle {
  id: number;
  licensePlate: string;
  make: string;
  model: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  vin?: string | null;
  /** @nullable */
  engineDisplacement?: number | null;
  /** @nullable */
  registrationDate?: string | null;
  /** @nullable */
  transmission?: VehicleTransmission;
  ownerType: VehicleOwnerType;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerAddress?: string | null;
  /** @nullable */
  ownerIco?: string | null;
  /** @nullable */
  ownerDic?: string | null;
  /** @nullable */
  ownerPhone?: string | null;
  /** @nullable */
  ownerEmail?: string | null;
  /** @nullable */
  consentGivenAt?: string | null;
  /** @nullable */
  consentNote?: string | null;
  /** @nullable */
  currentKm?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  stkValidUntil?: string | null;
  /** @nullable */
  lastOilChangeKm?: number | null;
  /** @nullable */
  lastOilChangeDate?: string | null;
  /** @nullable */
  lastBrakesDate?: string | null;
  /** @nullable */
  lastTimingDate?: string | null;
  /** @nullable */
  lastTransmissionOilDate?: string | null;
  /** @nullable */
  lastTransmissionOilKm?: number | null;
  /** @nullable */
  lastBrakeFluidDate?: string | null;
  /** @nullable */
  oilChangeIntervalKm?: number | null;
  /** @nullable */
  oilChangeIntervalMonths?: number | null;
  /** @nullable */
  transmissionOilIntervalKm?: number | null;
  /** @nullable */
  transmissionOilIntervalMonths?: number | null;
  /** @nullable */
  brakesIntervalMonths?: number | null;
  /** @nullable */
  timingIntervalKm?: number | null;
  /** @nullable */
  timingIntervalMonths?: number | null;
  /** @nullable */
  brakeFluidIntervalMonths?: number | null;
  createdAt: Date;
}